Made test helpers reusable
In addition to the ey-core library, other gems (notably our
internal tooling) more or less need access to up-to-date test
helpers.

This change moves these helpers from the non-distributed `spec`
dir to a distributable namespace, `Ey::Core::TestHelpers`.

Additionally, basic wiring for rspec/cucumber is achieved simply
by requiring either `ey-core/test_helpers/rspec` or
`ey-core/test_helpers/cucumber`, respectively.
