+++ lib/ey-core/test_helpers.rb
@@ -1,0 +1,15 @@
+require 'ey-core/test_helpers/account_helpers'
+require 'ey-core/test_helpers/alert_helpers'
+require 'ey-core/test_helpers/client_helpers'
+require 'ey-core/test_helpers/resource_helpers'
+
+module Ey
+  module Core
+    module TestHelpers
+      include AccountHelpers
+      include AlertHelpers
+      include ClientHelpers
+      include ResourceHelpers
+    end
+  end
+end
