+++ lib/ey-core/test_helpers/cucumber.rb
@@ -1,0 +1,3 @@
+require 'ey-core/test_helpers'
+
+World(Ey::Core::TestHelpers)
