+++ lib/ey-core/test_helpers/rspec.rb
@@ -1,0 +1,5 @@
+require 'ey-core/test_helpers'
+
+RSpec.configure do |config|
+  config.include(Ey::Core::TestHelpers)
+end
