+++ spec/support/helpers.rb
@@ -1,0 +1,1 @@
+require 'ey-core/test_helpers/rspec'
